requirements.txt
fastapi==0.68.1
h5py==3.1.0
imbalanced-learn==0.4.3
imblearn==0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
numpy==1.19.5
pandas==1.1.5
pandas-datareader==0.9.0
pandas-gbq==0.13.3
pandas-profiling==1.4.1
Pillow==7.1.2
pyngrok==5.1.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
scikit-learn==0.22.2.post1
scipy
sklearn==0.0
sklearn-pandas==1.8.0
tensorflow-cpu
uvicorn==0.13.1
uvloop==0.16.0
python-multipart